add a pausableAfterCount operator
The pause is available only after a certain count of events. Before the number of emitted events reaches that count
the Pauser will not be taken care of. When the Pauser is active, elements are ignored unless the second sequence
has most recently emitted `true`, this becomes the same behavior as the normal pausable operator.