maybe: add a ifEmpty(switchToSingleWithError:) extension
This commit adds an extension the trait "Maybe" that transforms it into a Single
whenever it is empty (ie completes without succeed).

The difference with the already existing "ifEmpty" function is that the Single
will immediately emmit the specified Error.

It allows to consider the case when the Maybe is empty as an applicative error.